r/aws_glue_trigger: Reduce visibility.
